Add separate module for popup
