+++ src/modules/render.js
@@ -1,15 +1,8 @@
-import showPopup from './showPopup.js';
-import Comment from './comments/commentAPI.js'; 
-import addComment from './comments/addComment.js';
-import {showComments, commentsContainer} from './comments/commentsBox.js'
-import commentHeader from './comments/commentHeader.js';
-import commentsCounter from './comments/commentsCounter.js';
 import '@fortawesome/fontawesome-free/js/fontawesome';
 import '@fortawesome/fontawesome-free/js/solid';
 import '@fortawesome/fontawesome-free/js/regular';
 import '@fortawesome/fontawesome-free/js/brands';
-
-const comment = new Comment();
+import showPopup from './popup';
 
 const LIKES_COUNT = 5;
 
@@ -71,45 +64,13 @@ const renderMainList = (ebookList) => {
     contentDiv.appendChild(commentButtonDiv);
 
     const popupSection = document.querySelector('.popup-section');
-    const popup = document.querySelector('.popup');
 
     const commentButton = document.createElement('button');
     commentButton.classList = 'contact-button';
     commentButton.addEventListener('click', async (e) => {
       e.preventDefault();
       popupSection.classList.toggle('hide');
-      popup.insertAdjacentHTML('beforeend', showPopup(ebook));
-      const crossIcon = document.querySelector('.cross-icon');
-      crossIcon.addEventListener('click', () => {
-        popupSection.classList.toggle('hide');
-        popup.innerHTML = '';
-      });
-
-      // display comments
-      const commentsArray = await comment.getComment(ebook.trackId);
-      popup.insertAdjacentHTML('beforeend', commentHeader());
-      popup.insertAdjacentHTML('beforeend', commentsContainer(commentsArray)); // displaying comments list
-
-      const commentList = document.querySelector('.comment-list');
-      const counterElement = document.querySelector('.comments-count');
-      counterElement.innerHTML = commentsCounter();
-
-      // add comments
-      popup.insertAdjacentHTML('beforeend', addComment());
-
-      const userName = document.getElementById('user-name');
-      const userComment = document.getElementById('user-comment');
-      const form = document.getElementById('comment-form');
-
-      form.addEventListener('submit', async (e) => {
-        e.preventDefault();
-        await comment.postComment(ebook.trackId, userName.value, userComment.value);
-        const commentsArray = await comment.getComment(ebook.trackId);
-        commentList.innerHTML = showComments(commentsArray);
-        counterElement.innerHTML = commentsCounter();
-        userName.value = '';
-        userComment.value = '';
-      });
+      showPopup(ebook);
     });
     commentButton.textContent = 'Comments';
     commentButtonDiv.appendChild(commentButton);
